requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.33.0
jinja2==3.1.6
python-dotenv==1.0.1

